requirements.txt
flask
pillow 
tensorflow
numpy
